Updated Spring Code

+++ src/main/java/com/fidelity/mts/service/AccountService.java
@@ -3,10 +3,12 @@ package com.fidelity.mts.service;
 import com.fidelity.mts.entity.Account;
 import com.fidelity.mts.entity.TransactionLog;
 
+import java.math.BigDecimal;
+
 public interface AccountService {
     Account getAccount(long id);
 
-    double getBalance(long id);
+    BigDecimal getBalance(long id);
 
     TransactionLog getTransactions(long id);
